+++ packages/botonic-plugin-ai-agents/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to Botonic will be documented in this file.
 ### Added
 
 - [PR-3059](https://github.com/hubtype/botonic/pull/3059): Add default tool `retrieve_knowledge`.
+- [PR-3109](https://github.com/hubtype/botonic/pull/3109): Add bot request in context for tools. This way you can use session and plugins in a function tool.
 
 ### Changed